# Backent/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EV Charging Platform API",
    "description": "API documentation for the EV Charging Platform project.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
...
    "/wallet/create": {
      "post": {
        "summary": "Create a new wallet",
        "operationId": "createWallet",
        "tags": ["Wallet"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Wallet created successfully"
          }
        }
      }
    },
    "/wallet/add": {
      "post": {
        "summary": "Add money to the wallet",
        "operationId": "addMoney",
        "tags": ["Wallet"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "format": "float"
                  }
                },
                "required": ["amount"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Money added to wallet"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 